+++ scripts/kish_basin_caverns_height.py
@@ -10,11 +10,9 @@ import contextily as cx
 import geopandas as gpd
 import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
-import numpy as np
 import pandas as pd
 import seaborn as sns
 from cartopy.mpl.ticker import LatitudeFormatter, LongitudeFormatter
-from matplotlib import ticker
 from matplotlib.lines import Line2D
 from matplotlib_scalebar.scalebar import ScaleBar
 
@@ -50,8 +48,7 @@ _, wells_b = fns.constraint_exploration_well(
 wind_farms = fns.constraint_wind_farm(
     data_path=os.path.join(
         "data", "wind-farms", "marine-area-consent-wind.zip"
-    ),
-    dat_extent=extent,
+    )
 )
 
 # frequent shipping routes
@@ -203,8 +200,11 @@ caverns[
     ]
 ].sum()
 
-# compare to Ireland's electricity demand in 2050 (Deane, 2021)
+# compare with Ireland's electricity demand in 2050 (Deane, 2021)
 compare.electricity_demand_ie(data=caverns["capacity"])
+
+# compare with Ireland's hydrogen demand in 2050
+compare.hydrogen_demand_ie(data=caverns["capacity"])
 
 # total capacity at various depth/height combinations
 s = caverns.groupby(["depth", "cavern_height", "halite"], sort=False)[
@@ -409,11 +409,11 @@ def plot_map_alt(dat_xr, cavern_df, zones_gdf, top_depth=True, fontsize=11.5):
     )
 
     plt.tight_layout()
-    plt.savefig(
-        os.path.join("graphics", "fig_caverns_capacity_ntg_height.jpg"),
-        format="jpg",
-        dpi=600,
-    )
+    # plt.savefig(
+    #     os.path.join("graphics", "fig_caverns_capacity_ntg_height.jpg"),
+    #     format="jpg",
+    #     dpi=600,
+    # )
     plt.show()