+++ scripts/wind_farm_optimisation.py
@@ -11,13 +11,10 @@ import geopandas as gpd
 import mapclassify as mc
 import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
 import seaborn as sns
 from cartopy.mpl.ticker import LatitudeFormatter, LongitudeFormatter
 from matplotlib import patheffects
 from matplotlib_scalebar.scalebar import ScaleBar
-from shapely.geometry import Point
 
 from h2ss import capacity as cap
 from h2ss import compare
@@ -50,8 +47,7 @@ _, wells_b = fns.constraint_exploration_well(
 wind_farms = fns.constraint_wind_farm(
     data_path=os.path.join(
         "data", "wind-farms", "marine-area-consent-wind.zip"
-    ),
-    dat_extent=extent,
+    )
 )
 
 # frequent shipping routes
@@ -119,6 +115,11 @@ caverns, _ = fns.generate_caverns_with_constraints(
 
 # ## Capacity
 
+print(
+    "Assumed density of hydrogen in the pipelines: "
+    f"{cap.HYDROGEN_DENSITY} kg m\N{SUPERSCRIPT MINUS}\N{SUPERSCRIPT THREE}"
+)
+
 caverns["cavern_total_volume"] = cap.cavern_volume(
     height=caverns["cavern_height"]
 )
@@ -168,7 +169,6 @@ weibull_wf_df = fns.read_weibull_data(
     data_path_wind_farms=os.path.join(
         "data", "wind-farms", "marine-area-consent-wind.zip"
     ),
-    dat_extent=extent,
 )
 
 weibull_powercurve = opt.weibull_distribution(weibull_wf_data=weibull_wf_df)
@@ -190,11 +190,9 @@ sns.despine()
 ax.xaxis.grid(True, linewidth=0.25)
 ax.yaxis.grid(True, linewidth=0.25)
 plt.tight_layout()
-plt.savefig(
-    os.path.join("graphics", "fig_powercurve.jpg"),
-    format="jpg",
-    dpi=600,
-)
+# plt.savefig(
+#     os.path.join("graphics", "fig_powercurve.jpg"), format="jpg", dpi=600
+# )
 plt.show()
 
 plt.figure(figsize=(10, 5.5))
@@ -220,11 +218,9 @@ ax.yaxis.grid(True, linewidth=0.25)
 sns.despine()
 ax.legend(title=None, fontsize=10.5)
 plt.tight_layout()
-plt.savefig(
-    os.path.join("graphics", "fig_weibull.jpg"),
-    format="jpg",
-    dpi=600,
-)
+# plt.savefig(
+#     os.path.join("graphics", "fig_weibull.jpg"), format="jpg", dpi=600
+# )
 plt.show()
 
 # ## Number of reference wind turbines
@@ -256,7 +252,7 @@ weibull_wf_df["demand"] = cap.energy_storage_capacity(
 
 # ## Number of caverns required based on cumulative working mass and AHP
 
-opt.calculate_number_of_caverns(
+compare.calculate_number_of_caverns(
     cavern_df=caverns, weibull_wf_data=weibull_wf_df
 )
 
@@ -285,11 +281,11 @@ weibull_wf_df[
 
 compare.electricity_demand_ie(data=weibull_wf_df["demand"])
 
+compare.hydrogen_demand_ie(data=weibull_wf_df["demand"])
+
 # ## LCOT for pipeline [€ kg⁻¹]
 
-caverns = opt.lcot_pipeline(
-    wf_data=wind_farms, weibull_wf_data=weibull_wf_df, cavern_df=caverns
-)
+caverns = opt.lcot_pipeline(weibull_wf_data=weibull_wf_df, cavern_df=caverns)
 
 caverns[
     [
@@ -341,11 +337,11 @@ axes[1].yaxis.grid(True, linewidth=0.25)
 axes[0].yaxis.grid(True, linewidth=0.25)
 sns.despine(bottom=True)
 plt.tight_layout()
-plt.savefig(
-    os.path.join("graphics", "fig_box_transmission_ntg.jpg"),
-    format="jpg",
-    dpi=600,
-)
+# plt.savefig(
+#     os.path.join("graphics", "fig_box_transmission_ntg.jpg"),
+#     format="jpg",
+#     dpi=600,
+# )
 plt.show()
 
 # ## Maps
@@ -356,7 +352,6 @@ shape = rd.halite_shape(dat_xr=ds).buffer(1000).buffer(-1000)
 def plot_map_facet(
     cavern_df,
     classes,
-    filename="fig_map_transmission_ntg.jpg",
     fontsize=11.5,
 ):
     """Helper function for plotting LCOT facet maps"""
@@ -438,13 +433,17 @@ def plot_map_facet(
         ax1.set_title(list(wind_farms["name"])[a])
 
     plt.tight_layout()
-    plt.savefig(os.path.join("graphics", filename), format="jpg", dpi=600)
+    # plt.savefig(
+    #     os.path.join("graphics", "fig_map_transmission_ntg.jpg"),
+    #     format="jpg", dpi=600
+    # )
     plt.show()
 
 
-classes = mc.Quantiles(caverns[list(caverns.filter(like="LCOT_"))], k=6)
-
-plot_map_facet(caverns, list(classes.bins))
+plot_map_facet(
+    caverns,
+    list(mc.Quantiles(caverns[list(caverns.filter(like="LCOT_"))], k=6).bins),
+)
 
 
 def plot_map_extent(cavern_df):
@@ -513,11 +512,11 @@ def plot_map_extent(cavern_df):
         )
     )
     plt.tight_layout()
-    plt.savefig(
-        os.path.join("graphics", "fig_transmission_map_inset.jpg"),
-        format="jpg",
-        dpi=600,
-    )
+    # plt.savefig(
+    #     os.path.join("graphics", "fig_transmission_map_inset.jpg"),
+    #     format="jpg",
+    #     dpi=600,
+    # )
     plt.show()