+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
-## 2024-04-28
+## [2024.04.2](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.04.2)
+
+[Various enhancements](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/11) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.04.1...2024.04.2).
+
+- fix and implement further tests (coverage at 63%)
+- regenerate all notebooks and rebuild docs
+- update docstrings
+- update tasks for linting / testing
+- update dataset metadata files
+- update CITATION.cff
+- workaround for injection point projection deprecation warning
+- reorganise / update functions and variable names
+- remove reassigned attributes from Xarray datasets
+
+## [2024.04.1](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.04.1)
+
+[Various enhancements](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/10) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.04.0...2024.04.1).
 
 - update task and script for converting notebooks
 - update documentation link - hosted on GitLab pages
@@ -25,3 +43,68 @@
 - remove old statistical plots of Kish basin data
 - update Kish basin halite net thickness facet maps
 - format notebooks and scripts
+
+## [2024.04.0](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.04.0)
+
+[Improvements to cavern calculations using net-to-gross](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/9) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.03.0...2024.04.0).
+
+## [2024.03.0](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.03.0)
+
+[Update docs](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/8) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.02.1...2024.03.0).
+
+## [2024.02.1](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.02.1)
+
+[Update plots](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/7) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.02.0...2024.02.1).
+
+## [2024.02.0](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.02.0)
+
+[Improved analysis and notebooks](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/6) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.6...2024.02.0).
+
+## [2024.01.6](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.6)
+
+[Update CAPEX/OPEX/electrolyser capacity calculations](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/5) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.5...2024.01.6).
+
+## [2024.01.5](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.5)
+
+[Finalise optimisation + update docs / plots](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/4) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.4...2024.01.5).
+
+## [2024.01.4](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.4)
+
+[Update package and docs](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/3) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.3...2024.01.4).
+
+## [2024.01.3](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.3)
+
+[Update configurations / metadata / docs](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/2) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.2...2024.01.3).
+
+## [2024.01.2](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.2)
+
+[Merge notebooks to main](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/pull/1) by @nmstreethran.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.1...2024.01.2).
+
+## [2024.01.1](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.1)
+
+Update optimisation / docs / tests and fixes.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2024.01.0...2024.01.1).
+
+## [2024.01.0](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2024.01.0)
+
+Initial docs / tests.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2023.12.1...2024.01.0).
+
+## [2023.12.1](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2022.12.1)
+
+Add optimisation functions.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/compare/2023.12.0...2023.12.1).
+
+## [2023.12.0](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/tree/2022.12.0)
+
+Initial version with capacity calculations.
+[**Full Changelog**](https://github.com/wind-to-hydrogen-toolkit/hydrogen-salt-storage/commits/2023.12.0).